move more help related stuff to help testing module
